Database builder/backend/requirements-minimal.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.10
langchain-community==0.0.35
pydantic==2.3.0
